+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingBodyContent/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingBodyContent, type IProposalVotingBodyContentProps } from './proposalVotingBodyContent';