+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingBodySummary/index.ts
@@ -1,0 +1,1 @@
+export { ProposalVotingBodySummary, type IProposalVotingBodySummaryProps } from './proposalVotingBodySummary';