+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingBodySummaryListItem/index.ts
@@ -1,0 +1,4 @@
+export {
+    ProposalVotingBodySummaryListItem,
+    type IProposalVotingBodySummaryListItemProps,
+} from './proposalVotingBodySummaryListItem';