+++ packages/gov-ui-kit/src/modules/components/proposal/proposalVoting/proposalVotingStageContext/proposalVotingStageContext.ts
@@ -1,7 +1,20 @@
 import { createContext, useContext } from 'react';
 import type { IProposalVotingStageProps } from '../proposalVotingStage/proposalVotingStage';
 
-export interface IProposalVotingStageContext extends Pick<IProposalVotingStageProps, 'startDate' | 'endDate'> {}
+export interface IProposalVotingStageContext extends Pick<IProposalVotingStageProps, 'startDate' | 'endDate'> {
+    /**
+     * List of plugin addresses to be displayed in the body summary list.
+     */
+    bodyList?: string[];
+    /**
+     * The active body to be displayed.
+     */
+    activeBody?: string;
+    /**
+     * Function to set the active body.
+     */
+    setActiveBody?: (id: string | undefined) => void;
+}
 
 const proposalVotingStageContext = createContext<IProposalVotingStageContext | null>(null);