+++ packages/gov-ui-kit/CHANGELOG.md
@@ -7,6 +7,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+-   Add new `ProposalVoting` module components (`ProposalVotingBodyContent`, `ProposalVotingBodySummary`,
+    `ProposalVotingBodySummaryList` and `ProposalVotingBodySummaryListItem`)
+-   Add story and export `ProposalVotingProgress` module component
+
+### Changed
+
+-   Update `ProposalVotingStage` module component to accept new optional `bodyList` prop and update rendering logic to
+    handle multi body proposals per stage.
+
 ## [1.0.56] - 2024-11-26
 
 ### Added